This PR updates projection loop in `DefaultSource` to use ordinals instead of field names. Also adds benchmark on projection.

Before:
```
Java HotSpot(TM) 64-Bit Server VM 1.7.0_80-b15 on Mac OS X 10.12.4
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
SQL project:                             Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
Riff (all fields)                              701 /  785          1.4         700.8       1.0X
Riff (1 field)                                 445 /  477          2.2         445.2       1.6X
Riff (3 fields)                                489 /  546          2.0         488.7       1.4X
Riff (6 fields)                                715 /  734          1.4         714.8       1.0X
```

After:
```
Java HotSpot(TM) 64-Bit Server VM 1.7.0_80-b15 on Mac OS X 10.12.4
Intel(R) Core(TM) i5-4258U CPU @ 2.40GHz
SQL project:                             Best/Avg Time(ms)    Rate(M/s)   Per Row(ns)   Relative
------------------------------------------------------------------------------------------------
Riff (all fields)                              723 /  794          1.4         722.7       1.0X
Riff (1 field)                                 435 /  463          2.3         435.0       1.7X
Riff (3 fields)                                468 /  506          2.1         467.7       1.5X
Riff (6 fields)                                676 /  684          1.5         676.4       1.1X
```